Change index type

+++ src/CrossPointState.h
@@ -9,7 +9,7 @@ class CrossPointState {
 
  public:
   std::string openEpubPath;
-  uint8_t lastSleepImage = UINT8_MAX;  // UINT8_MAX = unset sentinel
+  size_t lastSleepImage = SIZE_MAX;  // SIZE_MAX = unset sentinel
   uint8_t readerActivityLoadCount = 0;
   bool lastSleepFromReader = false;
   ~CrossPointState() = default;
